backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.11
botocore==1.34.11
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.9.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-nextjs==2.4.0
django-ses==3.5.2
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
lxml==5.0.1
Markdown==3.5.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
premailer==3.10.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
websockets==12.0
yagmail==0.15.293
yarl==1.9.4
